Move http and jetstream helpers
