+++ src/frontend/packages/store/src/entity-request-pipeline/entity-request-base-handlers/handle-multi-endpoints.pipe.ts
@@ -1,8 +1,8 @@
-import { hasJetStreamError, JetStreamErrorResponse } from '../../../../core/src/jetstream.helpers';
+import { NonJetstreamRequestHandler } from '../../entity-catalog/entity-catalog.types';
+import { hasJetStreamError, JetStreamErrorResponse } from '../../jetstream';
 import { PagedJetstreamResponse } from '../entity-request-pipeline.types';
 import { PaginationPageIteratorConfig } from '../pagination-request-base-handlers/pagination-iterator.pipe';
 import { stratosEndpointGuidKey } from '../pipeline.types';
-import { NonJetstreamRequestHandler } from '../../entity-catalog/entity-catalog.types';
 
 /**
  * Generic container for information about an errored request to a specific endpoint
